extract chat component from page so server action can be passed correctly
