+++ package.json
@@ -18,7 +18,7 @@
     "@supabase/supabase-js": "^2.38.4",
     "@types/nodegit": "^0.28.7",
     "@vercel/analytics": "^1.1.1",
-    "ai": "3.0.7",
+    "ai": "^2.2.31",
     "bufferutil": "^4.0.8",
     "copy-to-clipboard": "^3.3.3",
     "glob": "^10.3.3",
